Address review comments
