+++ java-frontend/src/main/java/org/sonar/java/model/springcontext/BeanDefinitionGatherer.java
@@ -16,22 +16,18 @@
  */
 package org.sonar.java.model.springcontext;
 
-import java.beans.Introspector;
-import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Base64;
 import java.util.LinkedHashMap;
 import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Collectors;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import javax.annotation.Nullable;
 import org.sonar.api.batch.fs.InputFile;
+import org.sonar.java.model.JUtils;
 import org.sonar.java.model.springcontext.TypeToDependenciesIndex.InjectionPoint;
 import org.sonar.java.reporting.AnalyzerMessage;
 import org.sonar.java.utils.PackageUtils;
@@ -39,9 +35,7 @@ import org.sonar.java.utils.SpringUtils;
 import org.sonar.plugins.java.api.InputFileScannerContext;
 import org.sonar.plugins.java.api.JavaFileScannerContext;
 import org.sonar.plugins.java.api.ModuleScannerContext;
-import org.sonar.plugins.java.api.semantic.Symbol;
 import org.sonar.plugins.java.api.semantic.SymbolMetadata;
-import org.sonar.plugins.java.api.semantic.Type;
 import org.sonar.plugins.java.api.tree.ClassTree;
 import org.sonar.plugins.java.api.tree.MethodTree;
 import org.sonar.plugins.java.api.tree.Tree;
@@ -76,24 +70,14 @@ public class BeanDefinitionGatherer extends SpringContextModelGatherer {
 
   private static final Logger LOG = LoggerFactory.getLogger(BeanDefinitionGatherer.class);
 
-  private static final String CACHE_KEY_PREFIX = "java:spring:bean-definitions:";
-  private static final String BEAN_SEPARATOR = "\n";
-  private static final String FIELD_SEPARATOR = "|";
-  private static final String DEP_SEPARATOR = ",";
-  private static final String DEP_KEY_VALUE_SEPARATOR = ":";
-  private static final String DEP_NAMES_SEPARATOR = ";";
-  private static final String DEP_LOCATION_SEPARATOR = "#";
-  private static final String TYPE_HIERARCHY_SEPARATOR = ";";
-
   private static final String PRIMARY_ANNOTATION = "org.springframework.context.annotation.Primary";
-  private static final String VALUE_ATTRIBUTE = "value";
 
   private final List<BeanData> collectedBeans = new ArrayList<>();
 
   /** Beans found in the file currently being scanned, used for per-file cache writes. */
   private final List<BeanData> beansCollectedAtFileLevel = new ArrayList<>();
 
-  private record BeanData(
+  record BeanData(
     String beanName,
     String type,
     String beanPackage,
@@ -117,9 +101,13 @@ public class BeanDefinitionGatherer extends SpringContextModelGatherer {
   }
 
   /**
-   * Visits class nodes and registers a bean when the class carries a stereotype annotation ({@code @Component},
+   * Visits class nodes and registers all beans defined in the class.
+   *
+   * Registers a bean when the class carries a stereotype annotation ({@code @Component},
    * {@code @Service}, {@code @Repository}, {@code @Controller}, {@code @RestController}, {@code @Configuration}),
-   * then looks for {@code @Bean} factory methods on that same class.
+   * then registers beans for {@code @Bean} factory methods on that same class.
+   *
+   * @param tree The class tree to visit
    */
   @Override
   public void visitNode(Tree tree) {
@@ -133,12 +121,12 @@ public class BeanDefinitionGatherer extends SpringContextModelGatherer {
     String pkg = PackageUtils.packageNameOf(classTree.symbol());
 
     if (SpringUtils.STEREOTYPE_ANNOTATIONS.stream().anyMatch(meta::isAnnotatedWith)) {
-      String beanName = extractBeanName(meta).orElseGet(() -> defaultBeanName(classTree.simpleName().name()));
+      String beanName = SpringUtils.extractBeanName(meta, classTree.simpleName().name());
       // collect autowired dependencies as InjectionPoints to store in TypeToDependenciesIndex
       Map<String, Set<InjectionPoint>> injectionPoints = collectAutowiredDependencies(classTree, context.getInputFile());
       // also collect their names mapped by type to store in dependingBeans
       Map<String, Set<String>> deps = toNameMap(injectionPoints);
-      Set<String> typeHierarchy = collectTypeHierarchy(classTree.symbol());
+      Set<String> typeHierarchy = JUtils.collectTypeHierarchy(classTree.symbol());
       var beanData = new BeanData(
         beanName, fqn, pkg,
         context.getInputFile(),
@@ -159,59 +147,20 @@ public class BeanDefinitionGatherer extends SpringContextModelGatherer {
   @Override
   public void leaveFile(JavaFileScannerContext context) {
     if (context.getCacheContext().isCacheEnabled()) {
-      writeToCache(context, beansCollectedAtFileLevel);
+      SpringContextCacheHelper.writeBeanDefinitionsToCache(context, LOG, beansCollectedAtFileLevel);
     }
     beansCollectedAtFileLevel.clear();
   }
 
-  /** Serializes and writes this file's beans as newline-joined lines, mirrored by {@link #deserializeBean} via {@code String#lines}. */
-  private static void writeToCache(JavaFileScannerContext context, List<BeanData> beans) {
-    var cacheKey = SpringContextCacheHelper.cacheKey(CACHE_KEY_PREFIX, context);
-    var data = beans.stream()
-      .map(BeanDefinitionGatherer::serializeBean)
-      .collect(Collectors.joining(BEAN_SEPARATOR));
-    SpringContextCacheHelper.writeToCache(context, LOG, cacheKey, data);
-  }
-
   /**
-   * Serializes one bean into a single "|"-delimited line. Any string sourced from user code (bean name,
-   * dependency type keys, injection point names) is Base64-encoded first, since {@code |}, {@code :},
-   * {@code ,}, {@code ;} or {@code #} could otherwise appear in an identifier and be mistaken for a
-   * field/entry separator.
-   */
-  private static String serializeBean(BeanData bean) {
-    var deps = bean.dependencyInjectionPoints().entrySet().stream()
-      .map(e -> Base64.getEncoder().encodeToString(e.getKey().getBytes(StandardCharsets.UTF_8))
-        + DEP_KEY_VALUE_SEPARATOR
-        + e.getValue().stream()
-          .map(BeanDefinitionGatherer::encodeInjectionPoint)
-          .collect(Collectors.joining(DEP_NAMES_SEPARATOR)))
-      .collect(Collectors.joining(DEP_SEPARATOR));
-    var typeHierarchy = String.join(TYPE_HIERARCHY_SEPARATOR, bean.typeHierarchy());
-    var span = bean.textSpan();
-    var encodedName = Base64.getEncoder().encodeToString(bean.beanName().getBytes(StandardCharsets.UTF_8));
-    return String.join(FIELD_SEPARATOR,
-      encodedName,
-      bean.type(),
-      bean.beanPackage(),
-      span.startLine + ":" + span.startCharacter + ":" + span.endLine + ":" + span.endCharacter,
-      Boolean.toString(bean.isPrimary()),
-      deps,
-      typeHierarchy);
-  }
-
-  /** Encodes one injection point (dependency name, location) as {@code <base64 name>#<span>}; see {@link #serializeBean} for why the name is encoded. */
-  private static String encodeInjectionPoint(InjectionPoint point) {
-    var span = point.location().mainLocation();
-    return Base64.getEncoder().encodeToString(point.name().getBytes(StandardCharsets.UTF_8))
-      + DEP_LOCATION_SEPARATOR
-      + span.startLine + ":" + span.startCharacter + ":" + span.endLine + ":" + span.endCharacter;
-  }
-
-  /**
-   * Transfers all beans collected across the module into the shared {@link SpringContextModel}: the bean
-   * definition itself, its position in every ancestor/interface type (for {@link TypeToBeanNamesIndex}), and
-   * each of its dependencies by type (for {@link TypeToDependenciesIndex}).
+   * Transfers all beans collected across the module into the shared {@link SpringContextModel}.
+   *
+   * Registers all encountered bean definitions in {@link BeanDefinitionRegistry},
+   * their position in every ancestor/interface type in {@link TypeToBeanNamesIndex}, and
+   * each of their dependencies by type in {@link TypeToDependenciesIndex}.
+   *
+   * @param context Scanner context used here to access the current module key
+   * @param springContextModel Shared cross-module Spring context
    */
   @Override
   public void gatherSpringContextData(ModuleScannerContext context, SpringContextModel springContextModel) {
@@ -235,112 +184,30 @@ public class BeanDefinitionGatherer extends SpringContextModelGatherer {
 
   @Override
   public boolean scanWithoutParsing(InputFileScannerContext ctx) {
-    return readFromCache(ctx).map(beans -> {
+    return SpringContextCacheHelper.readBeanDefinitionsFromCache(ctx, LOG).map(beans -> {
       collectedBeans.addAll(beans);
       return true;
     }).orElse(false);
   }
 
-  private static Optional<List<BeanData>> readFromCache(InputFileScannerContext ctx) {
-    var cacheKey = SpringContextCacheHelper.cacheKey(CACHE_KEY_PREFIX, ctx);
-    // A file with zero beans still has a cache entry (an empty string), handled explicitly here rather than
-    // relying on String#lines() returning an empty stream for "".
-    return SpringContextCacheHelper.readFromCache(ctx, LOG, cacheKey, content -> content.isEmpty()
-      ? List.<BeanData>of()
-      : content.lines().map(line -> deserializeBean(line, ctx.getInputFile())).toList());
-  }
-
   /**
-   * Reverse of {@link #serializeBean}: splits the "|"-delimited line back into a bean's fields, decoding
-   * every value that was Base64-encoded on write.
-   */
-  private static BeanData deserializeBean(String line, InputFile inputFile) {
-    // -1 keeps trailing empty fields (e.g. no dependencies/no type hierarchy) so the fixed field indices below stay aligned.
-    String[] fields = line.split("\\" + FIELD_SEPARATOR, -1);
-    String beanName = new String(Base64.getDecoder().decode(fields[0]), StandardCharsets.UTF_8);
-    String type = fields[1];
-    String beanPackage = fields[2];
-    String[] spanParts = fields[3].split(":");
-    var textSpan = new AnalyzerMessage.TextSpan(
-      Integer.parseInt(spanParts[0]),
-      Integer.parseInt(spanParts[1]),
-      Integer.parseInt(spanParts[2]),
-      Integer.parseInt(spanParts[3]));
-    boolean isPrimary = Boolean.parseBoolean(fields[4]);
-    Map<String, Set<InjectionPoint>> injectionPoints = new LinkedHashMap<>();
-    if (!fields[5].isEmpty()) {
-      for (String entry : fields[5].split(DEP_SEPARATOR)) {
-        // indexOf is safe: Base64 output never contains ':', so the first ':' is unambiguously the key/value boundary.
-        int idx = entry.indexOf(DEP_KEY_VALUE_SEPARATOR);
-        String typeFqn = new String(Base64.getDecoder().decode(entry.substring(0, idx)), StandardCharsets.UTF_8);
-        Set<InjectionPoint> points = Arrays.stream(entry.substring(idx + 1).split(DEP_NAMES_SEPARATOR))
-          .map(token -> decodeInjectionPoint(token, inputFile))
-          .collect(Collectors.toCollection(LinkedHashSet::new));
-        injectionPoints.put(typeFqn, points);
-      }
-    }
-    Map<String, Set<String>> deps = toNameMap(injectionPoints);
-    Set<String> typeHierarchy = !fields[6].isEmpty()
-      ? new LinkedHashSet<>(List.of(fields[6].split(TYPE_HIERARCHY_SEPARATOR)))
-      : new LinkedHashSet<>();
-    return new BeanData(beanName, type, beanPackage, inputFile, textSpan, isPrimary, deps, injectionPoints, typeHierarchy);
-  }
-
-  /** Reverse of {@link #encodeInjectionPoint}. {@code idx} lookup is safe since Base64 output never contains '#'. */
-  private static InjectionPoint decodeInjectionPoint(String token, InputFile inputFile) {
-    int idx = token.indexOf(DEP_LOCATION_SEPARATOR);
-    String name = new String(Base64.getDecoder().decode(token.substring(0, idx)), StandardCharsets.UTF_8);
-    String[] spanParts = token.substring(idx + 1).split(":");
-    var span = new AnalyzerMessage.TextSpan(
-      Integer.parseInt(spanParts[0]),
-      Integer.parseInt(spanParts[1]),
-      Integer.parseInt(spanParts[2]),
-      Integer.parseInt(spanParts[3]));
-    return new InjectionPoint(name, new BeanLocation(inputFile, span));
-  }
-
-  /**
-   * Finds an explicit bean name from whichever stereotype annotation is present, checking both the "value"
-   * attribute (e.g. {@code @Component("foo")}) and the "name" attribute (e.g. {@code @Controller(name = "foo")}).
-   * A class only ever carries one stereotype annotation in valid code, but
-   * this loops over all of them defensively rather than assuming which one is present.
-   */
-  private static Optional<String> extractBeanName(SymbolMetadata meta) {
-    for (String annotation : SpringUtils.STEREOTYPE_ANNOTATIONS) {
-      List<SymbolMetadata.AnnotationValue> attrs = meta.valuesForAnnotation(annotation);
-      if (attrs != null) {
-        Optional<String> name = attrs.stream()
-          .filter(v -> VALUE_ATTRIBUTE.equals(v.name()) || "name".equals(v.name()))
-          .map(v -> (String) v.value())
-          .filter(s -> !s.isBlank())
-          .findFirst();
-        if (name.isPresent()) {
-          return name;
-        }
-      }
-    }
-    return Optional.empty();
-  }
-
-  /** Spring's own default name for an unnamed bean: the class simple name with a decapitalized first letter. */
-  private static String defaultBeanName(String simpleName) {
-    return Introspector.decapitalize(simpleName);
-  }
-
-  /**
-   * Registers one {@link BeanData} per name declared on the {@code @Bean} method. In case of aliases like
-   * {@code @Bean({"a", "b"})} each name is recorded separately in the registry.
-   * Falls back to the method name when no name/value attribute is given, matching Spring's own default.
+   * Collects {@link BeanData} for a bean registered with the {@code @Bean} factory method.
+   *
+   * If multiple aliases are declared (e.g. {@code @Bean({"a", "b"})}), one {@link BeanData} is
+   * registered for each alias.
+   *
+   * @param method The {@code @Bean} factory method to visit
+   * @param pkg The bean's package (carried through to be stored in BeanData)
    */
   private void collectBeanMethod(MethodTree method, String pkg) {
     SymbolMetadata beanMeta = method.symbol().metadata();
-    List<String> beanNames = extractBeanMethodNames(beanMeta, method);
+    List<String> beanNames = SpringUtils.extractBeanMethodNames(beanMeta, method);
 
     String returnTypeFqn = method.returnType() != null
       ? method.returnType().symbolType().fullyQualifiedName()
       : "";
     Set<String> typeHierarchy = method.returnType() != null
-      ? collectTypeHierarchy(method.returnType().symbolType().symbol())
+      ? JUtils.collectTypeHierarchy(method.returnType().symbolType().symbol())
       : Set.of();
 
     var inputFile = context.getInputFile();
@@ -358,39 +225,16 @@ public class BeanDefinitionGatherer extends SpringContextModelGatherer {
   }
 
   /**
-   * Reads the {@code @Bean} method's explicit "value"/"name" attribute (accepting one or several aliases),
-   * falling back to the method's own name when none is given, matching Spring's own default.
-   */
-  private static List<String> extractBeanMethodNames(SymbolMetadata beanMeta, MethodTree method) {
-    List<SymbolMetadata.AnnotationValue> attrs = beanMeta.valuesForAnnotation(SpringUtils.BEAN_ANNOTATION);
-    List<String> names = new ArrayList<>();
-    if (attrs != null) {
-      for (SymbolMetadata.AnnotationValue attr : attrs) {
-        collectBeanMethodAttributeNames(attr, names);
-      }
-    }
-    if (names.isEmpty()) {
-      names.add(method.simpleName().name());
-    }
-    return names;
-  }
-
-  private static void collectBeanMethodAttributeNames(SymbolMetadata.AnnotationValue attr, List<String> names) {
-    if ((VALUE_ATTRIBUTE.equals(attr.name()) || "name".equals(attr.name())) && attr.value() instanceof Object[] values) {
-      for (Object value : values) {
-        if (value instanceof String name && !name.isBlank()) {
-          names.add(name);
-        }
-      }
-    }
-  }
-
-  /**
-   * Collects a class-level bean's dependencies from @Autowired fields/setters/constructors, plus Spring's
-   * implicit single-constructor injection: if no constructor is @Autowired and the class has exactly one
-   * constructor, Spring uses it for injection without requiring the annotation. {@code hasAutowiredConstructor}
-   * guards against misapplying that fallback when an @Autowired constructor already exists alongside other,
+   * Collects a class-level bean's dependencies from {@code @Autowired} fields, constructors and setters.
+   *
+   * Also applies Spring's implicit single-constructor injection if no constructor is {@code @Autowired}
+   * and the class declares exactly one constructor. {@code hasAutowiredConstructor} guards against
+   * misapplying that fallback when an {@code @Autowired} constructor already exists alongside other,
    * unannotated ones.
+   *
+   * @param classTree The class whose members are scanned for dependencies
+   * @param inputFile The file {@code classTree} was parsed from, used to locate each injection point
+   * @return The class's dependencies, mapped by required type FQN to the {@link InjectionPoint}s that require it
    */
   private static Map<String, Set<InjectionPoint>> collectAutowiredDependencies(ClassTree classTree, InputFile inputFile) {
     Map<String, Set<InjectionPoint>> deps = new LinkedHashMap<>();
@@ -399,7 +243,7 @@ public class BeanDefinitionGatherer extends SpringContextModelGatherer {
     for (Tree member : classTree.members()) {
       if (member instanceof VariableTree field && field.symbol().metadata().isAnnotatedWith(SpringUtils.AUTOWIRED_ANNOTATION)) {
         String typeFqn = field.symbol().type().fullyQualifiedName();
-        String name = dependencyKey(field.simpleName().name(), extractQualifier(field.symbol().metadata()));
+        String name = dependencyKey(field.simpleName().name(), SpringUtils.extractQualifier(field.symbol().metadata()));
         var location = new BeanLocation(inputFile, AnalyzerMessage.textSpanFor(field.simpleName()));
         deps.computeIfAbsent(typeFqn, k -> new LinkedHashSet<>()).add(new InjectionPoint(name, location));
       } else if (member instanceof MethodTree method) {
@@ -420,16 +264,18 @@ public class BeanDefinitionGatherer extends SpringContextModelGatherer {
   }
 
   /**
-   * Returns the given method's parameters as injection points, grouped by type FQN: each parameter becomes one
-   * {@link InjectionPoint} whose dependency name is its {@code @Qualifier} value if present, otherwise its own
-   * name. Shared by {@code @Autowired} constructors/setters and {@code @Bean} factory methods, whose
-   * dependencies both come from method parameters.
+   * Collect the given method's parameters as dependencies.
+   *
+   * @param method Method whose parameters are stored as dependencies, either {@code @Autowired} constructors/setters or
+   * {@code @Bean} factory methods
+   * @param inputFile The file {@code method} was parsed from, used to locate each injection point
+   * @return The collected dependencies, mapped by required type FQN to the {@link InjectionPoint}s that require it
    */
   private static Map<String, Set<InjectionPoint>> parameterDependencies(MethodTree method, InputFile inputFile) {
     Map<String, Set<InjectionPoint>> deps = new LinkedHashMap<>();
     for (var p : method.parameters()) {
       String typeFqn = p.symbol().type().fullyQualifiedName();
-      String name = dependencyKey(p.simpleName().name(), extractQualifier(p.symbol().metadata()));
+      String name = dependencyKey(p.simpleName().name(), SpringUtils.extractQualifier(p.symbol().metadata()));
       var location = new BeanLocation(inputFile, AnalyzerMessage.textSpanFor(p.simpleName()));
       deps.computeIfAbsent(typeFqn, k -> new LinkedHashSet<>()).add(new InjectionPoint(name, location));
     }
@@ -437,7 +283,7 @@ public class BeanDefinitionGatherer extends SpringContextModelGatherer {
   }
 
   /** Projects each type's injection points down to just their names, discarding location — the flat view stored in {@code BeanDefinitionHolder}. */
-  private static Map<String, Set<String>> toNameMap(Map<String, Set<InjectionPoint>> injectionPointsByType) {
+  static Map<String, Set<String>> toNameMap(Map<String, Set<InjectionPoint>> injectionPointsByType) {
     Map<String, Set<String>> names = new LinkedHashMap<>();
     injectionPointsByType.forEach((typeFqn, points) -> names.put(typeFqn, points.stream()
       .map(InjectionPoint::name)
@@ -445,57 +291,8 @@ public class BeanDefinitionGatherer extends SpringContextModelGatherer {
     return names;
   }
 
-  /**
-   * The name Spring uses to resolve this dependency: an explicit @Qualifier
-   * value if present, otherwise the field/parameter name.
-   */
   private static String dependencyKey(String fieldOrParamName, @Nullable String qualifier) {
     return qualifier != null ? qualifier : fieldOrParamName;
-  }
-
-  /**
-   * Reads the {@code @Qualifier} value, if any. A missing or blank value is treated as no
-   * qualifier, so callers fall back to the field/parameter name via {@link #dependencyKey}.
-   */
-  @Nullable
-  private static String extractQualifier(SymbolMetadata metadata) {
-    List<SymbolMetadata.AnnotationValue> attrs = metadata.valuesForAnnotation(SpringUtils.QUALIFIER_ANNOTATION);
-    if (attrs == null) {
-      return null;
-    }
-    return attrs.stream()
-      .filter(v -> VALUE_ATTRIBUTE.equals(v.name()))
-      .map(v -> (String) v.value())
-      .filter(s -> !s.isBlank())
-      .findFirst()
-      .orElse(null);
-  }
-
-  private static Set<String> collectTypeHierarchy(Symbol.TypeSymbol symbol) {
-    Set<String> visited = new LinkedHashSet<>();
-    walkTypeHierarchy(symbol, visited);
-    return visited;
-  }
-
-  /**
-   * Recursively walks superclasses and interfaces, recording each type's FQN. {@code java.lang.Object} and
-   * unknown/unresolved types are excluded, and {@code visited} serves as both the result and a guard against
-   * walking the same type twice, e.g. a common ancestor reached through more than one interface.
-   */
-  private static void walkTypeHierarchy(Symbol.TypeSymbol symbol, Set<String> visited) {
-    String fqn = symbol.type().fullyQualifiedName();
-    if ("java.lang.Object".equals(fqn) || symbol.type().isUnknown() || !visited.add(fqn)) {
-      return;
-    }
-    Type superClass = symbol.superClass();
-    if (superClass != null && !superClass.isUnknown()) {
-      walkTypeHierarchy(superClass.symbol(), visited);
-    }
-    for (Type iface : symbol.interfaces()) {
-      if (!iface.isUnknown()) {
-        walkTypeHierarchy(iface.symbol(), visited);
-      }
-    }
   }
 
 }
